requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# flask-argon-dashboard/app/__init__.py: 9
# flask-argon-dashboard/app/views.py: 13
Flask == 1.1.1

# flask-argon-dashboard/app/__init__.py: 12
Flask_Bcrypt == 0.7.1

# flask-argon-dashboard/app/views.py: 9
Flask_Jsonpify == 1.5.0

# flask-argon-dashboard/app/__init__.py: 11
# flask-argon-dashboard/app/models.py: 8
# flask-argon-dashboard/app/views.py: 14
Flask_Login == 0.5.0

# flask-argon-dashboard/app/__init__.py: 10
Flask_SQLAlchemy == 2.4.1

# flask-argon-dashboard/app/forms.py: 8
Flask_WTF == 0.14.3

# flask-argon-dashboard/app/engine.py: 1,2,3,4
Keras == 2.3.1

# flask-argon-dashboard/app/csv_to_db.py: 4
# flask-argon-dashboard/app/db_to_df.py: 4
SQLAlchemy == 1.3.9

# flask-argon-dashboard/app/forms.py: 10
WTForms == 2.3.1

# flask-argon-dashboard/app/views.py: 15
Werkzeug == 0.16.0

# flask-argon-dashboard/app/views.py: 10
folium == 0.10.1

# flask-argon-dashboard/app/engine.py: 7
numpy == 1.17.2

# flask-argon-dashboard/app/csv_to_db.py: 1
# flask-argon-dashboard/app/db_to_df.py: 1
# flask-argon-dashboard/app/engine.py: 10
pandas == 0.24.2

# flask-argon-dashboard/app/engine.py: 9
python_dateutil == 2.7.5

# flask-argon-dashboard/app/engine.py: 6
scikit_learn == 0.21.3
